Feat/multimedia inmueble crud

+++ AirbnbUdc.Repository.Implementation/DataModel/PropertyOwner.cs
@@ -20,7 +20,7 @@ namespace AirbnbUdc.Repository.Implementation.DataModel
             this.Property = new HashSet<Property>();
         }
     
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string FirstName { get; set; }
         public string FamilyName { get; set; }
         public string Email { get; set; }
@@ -29,5 +29,7 @@ namespace AirbnbUdc.Repository.Implementation.DataModel
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Property> Property { get; set; }
+
+
     }
 }
